change datgbase

+++ models/personne.js
@@ -2,7 +2,7 @@
 module.exports = (sequelize, DataTypes) => {
   const Personne = sequelize.define("Personne", {
     cin_p: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.INTEGER(8),
       allowNull: false,
     },
   });
